+++ azure/rs_azure_compute/changelog.md
@@ -1,5 +1,14 @@
 Compute Plugin changelog
 
+v1.2 (08-01-2017)
+-----------------
+- fixing compute extensions to fail properly
+
+v1.1 (08-01-2017)
+-----------------
+- adding compute extensions
+- moving from vm_name to virtualMachineName
+
 v1.0
 -----
 - initial release